fixes for errors during training (empty dataset, None->tf.Tensor conversion) - overhauled EDGS to not use datasets anymore - in EDGM.__get_item__: adjusted lengths of placeholder vectors
outstanding problem in EDGS:
```
adj = tf.reshape(tf.stack(adj), [self.batch_size, 1])

Input to reshape is a tensor with 0 values, but the requested shape has
4 [Op:Reshape]

```

+++ tests/test_data_generation.py
@@ -285,7 +285,7 @@ class TestEdgeDGMultiple(unittest.TestCase):
         is_normalised = np.max(combo_img) <= 1
         self.assertTrue(is_normalised)
         self.assertEqual(combo_img.shape, (256, 256, 2))
-        self.assertEqual(combo_img.dtype, np.int64)
+        self.assertEqual(combo_img.dtype, np.int32)
 
         # visual test
         skel_part = combo_img[..., 0]
